Renamed file to authGithubSigninPopup
